requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.13.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
loguru==0.7.2
google-genai
langchain
langgraph
pandas==2.3.3
openpyxl==3.1.5
pymilvus
sentence-transformers
